+++ SudoDEM3D/gui/qt5/GLViewer.hpp
@@ -153,7 +153,7 @@ class SudoDEMCamera : public qglviewer::Camera
 		float cuttingDistance;
         public :
 		SudoDEMCamera():cuttingDistance(0){};
-		 qreal zNear() const;
+		 float zNear() const;
 		virtual void setCuttingDistance(float s){cuttingDistance=s;};
 };
